+++ Ai_plant_lab/src/AIPlantLab/Experiments.Application/IntegrationEvents/ExperimentConfiguredIntegrationEvent.cs
@@ -2,7 +2,7 @@ namespace Experiments.Application.IntegrationEvents;
 
 public sealed record ExperimentConfiguredIntegrationEvent(
     Guid ExperimentId,
-    int LightHoursPerDay,
+    decimal LightHoursPerDay,
     int WateringIntervalDays,
     decimal TargetTemperatureCelsius,
     string? Notes,
