+++ Ai_plant_lab/src/AIPlantLab/Experiments.Domain/ValueObjects/ExperimentConfiguration.cs
@@ -1,7 +1,7 @@
 namespace Experiments.Domain.ValueObjects;
 
 public sealed record ExperimentConfiguration(
-    int LightHoursPerDay,
+    decimal LightHoursPerDay,
     int WateringIntervalDays,
     decimal TargetTemperatureCelsius,
     string? Notes);
